@@ -16,7 +16,8 @@ def main(mode: Literal["naive", "hybrid"], query: str):
     custom_rag = CustomLightRAG(
         working_dir=working_dir,
         graph_storage="TigerGraphStorage",
-        vector_storage="NanoVectorDBStorage",
+        vector_storage="TigerVectorStorage",  # Use TigerGraph for storing vectors
+        # vector_storage="NanoVectorDBStorage",
         kv_storage="JsonKVStorage",
     )
 
