@@ -5,10 +5,12 @@ from .graph_db import (
     # configurations for graph schema
     DataType,
     AttributeSchema,
+    AttributesType,
+    VectorAttributeSchema,
+    VectorAttributesType,
     NodeSchema,
     EdgeSchema,
     GraphSchema,
-    AttributesType,
     create_node_schema,
     create_edge_schema,
     # configurations for loading job
@@ -31,7 +33,7 @@ from .settings import (
     BaseEmbeddingConfig,
     OpenAIEmbeddingConfig,
     BaseVectorDBConfig,
-    LanceDBConfig,
+    TigerVectorConfig,
     NanoVectorDBConfig,
     BaseChatConfig,
     OpenAIChatConfig,
@@ -45,10 +47,12 @@ __all__ = [
     # configurations for graph schema
     "DataType",
     "AttributeSchema",
+    "AttributesType",
+    "VectorAttributeSchema",
+    "VectorAttributesType",
     "NodeSchema",
     "EdgeSchema",
     "GraphSchema",
-    "AttributesType",
     "create_node_schema",
     "create_edge_schema",
     # configurations for loading job
@@ -70,7 +74,7 @@ __all__ = [
     "BaseEmbeddingConfig",
     "OpenAIEmbeddingConfig",
     "BaseVectorDBConfig",
-    "LanceDBConfig",
+    "TigerVectorConfig",
     "NanoVectorDBConfig",
     "BaseChatConfig",
     "OpenAIChatConfig",
