@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import List, Optional, Set
 from pydantic import Field
 
 from ..base_config import BaseConfig
@@ -11,12 +11,15 @@ class NeighborSpec(BaseConfig):
 
     start_nodes: str | List[str] = Field(..., description="List of starting node IDs.")
     start_node_type: str = Field(..., description="The type of the start node.")
-    edge_types: Optional[str | List[str]] = Field(
-        None, description="List of allowed edge types for traversal."
+    start_node_alias: str = Field("s", description="Alias for the source node.")
+    edge_type_set: Optional[Set[str]] = Field(
+        None, description="Set of allowed edge types for traversal."
     )
-    target_node_types: Optional[str | List[str]] = Field(
-        None, description="List of allowed target node types."
+    edge_alias: str = Field("e", description="Alias for the edge.")
+    target_node_type_set: Optional[Set[str]] = Field(
+        None, description="Set of allowed target node types."
     )
+    target_node_alias: str = Field("t", description="Alias for the target node.")
     filter_expression: Optional[str] = Field(
         None, description="A string defining complex filtering logic."
     )
