@@ -1,7 +1,7 @@
 from .settings import Settings
 from .llm_settings import BaseLLMConfig, OpenAIConfig
 from .embedding_settings import BaseEmbeddingConfig, OpenAIEmbeddingConfig
-from .vector_db_settings import BaseVectorDBConfig, LanceDBConfig, NanoVectorDBConfig
+from .vector_db_settings import BaseVectorDBConfig, TigerVectorConfig, NanoVectorDBConfig
 from .chat_settings import BaseChatConfig, OpenAIChatConfig
 
 __all__ = [
@@ -11,7 +11,7 @@ __all__ = [
     "BaseEmbeddingConfig",
     "OpenAIEmbeddingConfig",
     "BaseVectorDBConfig",
-    "LanceDBConfig",
+    "TigerVectorConfig",
     "NanoVectorDBConfig",
     "BaseChatConfig",
     "OpenAIChatConfig",
