@@ -1,5 +1,5 @@
 import logging
-from typing import Dict, List
+from typing import Any, Dict, List, Optional, Tuple
 
 from .base_manager import BaseManager
 
@@ -30,10 +30,29 @@ class EdgeManager(BaseManager):
             logger.error(f"Error adding from {src_node_id} to {tgt_node_id}: {e}")
             return None
 
+    def add_edges_from(
+        self,
+        normalized_edges: List[Tuple[str, str, Dict[str, Any]]],
+        src_node_type: str,
+        edge_type: str,
+        tgt_node_type: str,
+    ) -> Optional[int]:
+        try:
+            result = self._connection.upsertEdges(
+                sourceVertexType=src_node_type,
+                edgeType=edge_type,
+                targetVertexType=tgt_node_type,
+                edges=normalized_edges,
+            )
+            return result
+        except Exception as e:
+            logger.error(f"Error adding edges: {e}")
+            return None
+
     def has_edge(
         self,
-        src_node_id: str | int,
-        tgt_node_id: str | int,
+        src_node_id: str,
+        tgt_node_id: str,
         src_node_type: str,
         edge_type: str,
         tgt_node_type: str,
@@ -43,10 +62,7 @@ class EdgeManager(BaseManager):
                 src_node_type, src_node_id, edge_type, tgt_node_type, tgt_node_id
             )
             return bool(result)
-        except Exception as e:
-            # logger.error(
-            #     f"Error checking existence of edge from {src_node_id} to {tgt_node_id}: {e}"
-            # )
+        except Exception:
             return False
 
     def get_edge_data(
@@ -61,12 +77,9 @@ class EdgeManager(BaseManager):
             result = self._connection.getEdges(
                 src_node_type, src_node_id, edge_type, tgt_node_type, tgt_node_id
             )
-            if isinstance(result, List) and result:
-                return result[0].get("attributes", None)
+            if isinstance(result, List) and result:  # pyright: ignore
+                return result[0].get("attributes", None)  # pyright: ignore
             else:
                 raise TypeError(f"Unsupported type for result: {type(result)}")
-        except Exception as e:
-            # logger.error(
-            #     f"Error retrieving edge from {src_node_id} to {tgt_node_id}: {e}"
-            # )
+        except Exception:
             return None
