@@ -1,5 +1,5 @@
 import logging
-from typing import List, Dict, Optional, Tuple
+from typing import List, Dict, Optional, Set, Tuple
 import pandas as pd
 
 from tigergraphx.config import (
@@ -30,38 +30,42 @@ class QueryManager(BaseManager):
 
     def get_nodes(
         self,
-        node_type: str = "",
+        node_type: Optional[str] = None,
+        all_node_types: bool = False,
+        node_alias: str = "s",
         filter_expression: Optional[str] = None,
         return_attributes: Optional[str | List[str]] = None,
         limit: Optional[int] = None,
-    ) -> pd.DataFrame | None:
+    ) -> pd.DataFrame:
         """
         High-level function to retrieve nodes with multiple parameters.
         Converts parameters into a NodeSpec and delegates to `_get_nodes_from_spec`.
         """
         spec = NodeSpec(
             node_type=node_type,
+            all_node_types=all_node_types,
+            node_alias=node_alias,
             filter_expression=filter_expression,
             return_attributes=return_attributes,
             limit=limit,
         )
         return self.get_nodes_from_spec(spec)
 
-    def get_nodes_from_spec(self, spec: NodeSpec) -> pd.DataFrame | None:
+    def get_nodes_from_spec(self, spec: NodeSpec) -> pd.DataFrame:
         """
         Core function to retrieve nodes based on a NodeSpec object.
         """
-        gsql_script = self._create_gsql_get_nodes(self._graph_schema.graph_name, spec)
+        gsql_script = self._create_gsql_get_nodes(spec)
         try:
             result = self._connection.runInterpretedQuery(gsql_script)
             if not result or not isinstance(result, list):
-                return None
+                return pd.DataFrame()
             nodes = result[0].get("Nodes")
             if not nodes or not isinstance(nodes, list):
-                return None
+                return pd.DataFrame()
             df = pd.DataFrame(pd.json_normalize(nodes))
             if df.empty:
-                return None
+                return pd.DataFrame()
             attribute_columns = [
                 col for col in df.columns if col.startswith("attributes.")
             ]
@@ -92,18 +96,21 @@ class QueryManager(BaseManager):
             return pd.DataFrame(df[reordered_columns + remaining_columns])
         except Exception as e:
             logger.error(f"Error retrieving nodes for type {spec.node_type}: {e}")
-        return None
+        return pd.DataFrame()
 
     def get_neighbors(
         self,
         start_nodes: str | List[str],
         start_node_type: str,
-        edge_types: Optional[str | List[str]] = None,
-        target_node_types: Optional[str | List[str]] = None,
+        start_node_alias: str = "s",
+        edge_type_set: Optional[Set[str]] = None,
+        edge_alias: str = "e",
+        target_node_type_set: Optional[Set[str]] = None,
+        target_node_alias: str = "t",
         filter_expression: Optional[str] = None,
         return_attributes: Optional[str | List[str]] = None,
         limit: Optional[int] = None,
-    ) -> pd.DataFrame | None:
+    ) -> pd.DataFrame:
         """
         High-level function to retrieve neighbors with multiple parameters.
         Converts parameters into a NeighborSpec and delegates to `_get_neighbors_from_spec`.
@@ -111,31 +118,32 @@ class QueryManager(BaseManager):
         spec = NeighborSpec(
             start_nodes=start_nodes,
             start_node_type=start_node_type,
-            edge_types=edge_types,
-            target_node_types=target_node_types,
+            start_node_alias=start_node_alias,
+            edge_type_set=edge_type_set,
+            edge_alias=edge_alias,
+            target_node_type_set=target_node_type_set,
+            target_node_alias=target_node_alias,
             filter_expression=filter_expression,
             return_attributes=return_attributes,
             limit=limit,
         )
         return self.get_neighbors_from_spec(spec)
 
-    def get_neighbors_from_spec(self, spec: NeighborSpec) -> pd.DataFrame | None:
+    def get_neighbors_from_spec(self, spec: NeighborSpec) -> pd.DataFrame:
         """
         Core function to retrieve neighbors based on a NeighborSpec object.
         """
-        gsql_script, params = self._create_gsql_get_neighbors(
-            self._graph_schema.graph_name, spec
-        )
+        gsql_script, params = self._create_gsql_get_neighbors(spec)
         try:
             result = self._connection.runInterpretedQuery(gsql_script, params)
             if not result or not isinstance(result, list):
-                return None
+                return pd.DataFrame()
             neighbors = result[0].get("Neighbors")
             if not neighbors or not isinstance(neighbors, list):
-                return None
+                return pd.DataFrame()
             df = pd.DataFrame(pd.json_normalize(neighbors))
             if df.empty:
-                return None
+                return pd.DataFrame()
             attribute_columns = [
                 col for col in df.columns if col.startswith("attributes.")
             ]
@@ -164,14 +172,14 @@ class QueryManager(BaseManager):
             logger.error(
                 f"Error retrieving neighbors for node(s) {spec.start_nodes}: {e}"
             )
-        return None
+        return pd.DataFrame()
 
-    @staticmethod
-    def _create_gsql_get_nodes(graph_name: str, spec: NodeSpec) -> str:
+    def _create_gsql_get_nodes(self, spec: NodeSpec) -> str:
         """
         Core function to generate a GSQL query based on a NodeSpec object.
         """
-        node_type_str = f"{spec.node_type}.*" if spec.node_type else "ANY"
+        graph_name = self._graph_schema.graph_name
+        node_type_str = f"{spec.node_type}.*" if not spec.all_node_types else "ANY"
         filter_expression_str = (
             f"WHERE {spec.filter_expression}" if spec.filter_expression else ""
         )
@@ -185,9 +193,9 @@ INTERPRET QUERY() FOR GRAPH {graph_name} {{
 """
         # Add SELECT block only if filter or limit is specified
         if filter_expression_str or limit_clause:
-            query += """  Nodes =
-    SELECT s
-    FROM Nodes:s
+            query += f"""  Nodes =
+    SELECT {spec.node_alias}
+    FROM Nodes:{spec.node_alias}
 """
             if filter_expression_str:
                 query += f"    {filter_expression_str}\n"
@@ -207,14 +215,12 @@ INTERPRET QUERY() FOR GRAPH {graph_name} {{
         query += "\n}"
         return query.strip()
 
-    @staticmethod
-    def _create_gsql_get_neighbors(
-        graph_name: str, spec: NeighborSpec
-    ) -> Tuple[str, str]:
+    def _create_gsql_get_neighbors(self, spec: NeighborSpec) -> Tuple[str, str]:
         """
         Core function to generate a GSQL query based on a NeighborSpec object.
         """
         # Normalize fields to lists
+        graph_name = self._graph_schema.graph_name
         params = "&".join(
             [
                 f"start_nodes={node}"
@@ -224,14 +230,6 @@ INTERPRET QUERY() FOR GRAPH {graph_name} {{
                     else spec.start_nodes
                 )
             ]
-        )
-        edge_types = (
-            [spec.edge_types] if isinstance(spec.edge_types, str) else spec.edge_types
-        )
-        target_node_types = (
-            [spec.target_node_types]
-            if isinstance(spec.target_node_types, str)
-            else spec.target_node_types
         )
         return_attributes = (
             [spec.return_attributes]
@@ -248,16 +246,18 @@ INTERPRET QUERY() FOR GRAPH {graph_name} {{
         # Prepare components
         start_node_type = spec.start_node_type
         edge_types_str = (
-            f"(({ '|'.join(edge_types or []) }):e)"
-            if edge_types and len(edge_types) > 1
-            else f"({ '|'.join(edge_types or []) }:e)"
-            if edge_types
-            else "(:e)"
+            f"(({'|'.join(spec.edge_type_set)}):{spec.edge_alias})"
+            if spec.edge_type_set and len(spec.edge_type_set) > 1
+            else f"({'|'.join(spec.edge_type_set)}:{spec.edge_alias})"
+            if spec.edge_type_set is not None
+            else f"(:{spec.edge_alias})"
         )
         target_node_types_str = (
-            f"(({ '|'.join(target_node_types or []) }))"
-            if target_node_types and len(target_node_types) > 1
-            else f"{ '|'.join(target_node_types or []) }"
+            f"(({'|'.join(spec.target_node_type_set)}))"
+            if spec.target_node_type_set and len(spec.target_node_type_set) > 1
+            else f"{'|'.join(spec.target_node_type_set)}"
+            if spec.target_node_type_set is not None
+            else ""
         )
 
         where_clause = (
@@ -266,14 +266,16 @@ INTERPRET QUERY() FOR GRAPH {graph_name} {{
         limit_clause = f"    LIMIT {spec.limit}" if spec.limit else ""
 
         # Generate the query
+        s_alias = spec.start_node_alias
+        t_alias = spec.target_node_alias
         query = f"""
 INTERPRET QUERY(
   SET<VERTEX<{start_node_type}>> start_nodes
 ) FOR GRAPH {graph_name} {{
   Nodes = {{start_nodes}};
   Neighbors =
-    SELECT t
-    FROM Nodes:s -{edge_types_str}- {target_node_types_str}:t
+    SELECT {t_alias}
+    FROM Nodes:{s_alias} -{edge_types_str}- {target_node_types_str}:{t_alias}
 """
         if where_clause:
             query += f"{where_clause}\n"
