@@ -7,17 +7,20 @@ from tenacity import (
 
 
 class RetryMixin:
-    """Mixin for initializing a retry mechanism."""
+    """
+    Mixin for initializing a retry mechanism with configurable parameters.
+    """
 
     def initialize_retryer(self, max_retries: int, max_wait: int) -> AsyncRetrying:
-        """Initialize the retry mechanism.
+        """
+        Initialize the retry mechanism with exponential backoff and jitter.
 
         Args:
-            max_retries (int): Maximum number of retry attempts.
-            max_wait (int): Maximum wait time between retries in seconds.
+            max_retries: Maximum number of retry attempts.
+            max_wait: Maximum wait time between retries in seconds.
 
         Returns:
-            AsyncRetrying: Configured retrying instance.
+            AsyncRetrying: Configured retrying instance with specified parameters.
         """
         return AsyncRetrying(
             stop=stop_after_attempt(max_retries),
