@@ -1,14 +1,22 @@
 from .embedding import BaseEmbedding, OpenAIEmbedding
-from .vector_db import BaseVectorDB, LanceDBManager, NanoVectorDBManager
-from .search import BaseSearchEngine, LanceDBSearchEngine, NanoVectorDBSearchEngine
+from .vector_db import (
+    BaseVectorDB,
+    NanoVectorDBManager,
+    TigerVectorManager,
+)
+from .search import (
+    BaseSearchEngine,
+    TigerVectorSearchEngine,
+    NanoVectorDBSearchEngine,
+)
 
 __all__ = [
     "BaseEmbedding",
     "OpenAIEmbedding",
     "BaseVectorDB",
-    "LanceDBManager",
+    "TigerVectorManager",
     "NanoVectorDBManager",
     "BaseSearchEngine",
-    "LanceDBSearchEngine",
+    "TigerVectorSearchEngine",
     "NanoVectorDBSearchEngine",
 ]
