@@ -12,19 +12,19 @@ class BaseEmbedding(ABC):
         Initialize the base embedding model.
 
         Args:
-            config (BaseEmbeddingConfig): Configuration for the embedding model.
+            config: Configuration for the embedding model.
         """
         self.config = config
 
     @abstractmethod
     async def generate_embedding(self, text: str) -> List[float]:
         """
-        Asynchronously generate embeddings for a given text.
+        Asynchronously generate an embedding for the given text.
 
         Args:
-            text (str): The input text to generate embeddings for.
+            text: Input text to generate an embedding.
 
         Returns:
-            List[float]: A list of floats representing the text embedding.
+            A list of floats representing the text embedding.
         """
         pass
