@@ -1,9 +1,9 @@
 from .base_search_engine import BaseSearchEngine
-from .lancedb_search_engine import LanceDBSearchEngine
+from .tigervector_search_engine import TigerVectorSearchEngine
 from .nano_vectordb_search_engine import NanoVectorDBSearchEngine
 
 __all__ = [
     "BaseSearchEngine",
-    "LanceDBSearchEngine",
+    "TigerVectorSearchEngine",
     "NanoVectorDBSearchEngine",
 ]
