@@ -1,9 +1,9 @@
 from .base_vector_db import BaseVectorDB
-from .lancedb_manager import LanceDBManager
+from .tigervector_manager import TigerVectorManager
 from .nano_vectordb_manager import NanoVectorDBManager
 
 __all__ = [
     "BaseVectorDB",
-    "LanceDBManager",
+    "TigerVectorManager",
     "NanoVectorDBManager",
 ]
