@@ -1,6 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import Any, Dict, List
-from pathlib import Path
+from typing import List
 import pandas as pd
 
 from tigergraphx.config import BaseVectorDBConfig
@@ -11,55 +10,20 @@ class BaseVectorDB(ABC):
 
     def __init__(self, config: BaseVectorDBConfig):
         """
-        Initialize the base vector DB manager.
+        Initialize the vector database connection.
 
         Args:
-            config (BaseVectorDBConfig): Configuration for the vector database connection.
+            config: Configuration for the vector DB connection.
         """
         self.config = config
 
     @abstractmethod
-    def connect(self, uri: str | Path, **kwargs: Any) -> None:
-        """
-        Connect to the vector database and set up the connection.
-
-        Args:
-            uri (str | Path): The URI or path to connect to the vector database.
-            **kwargs (Any): Additional keyword arguments for the connection.
-        """
-        pass
-
-    @abstractmethod
-    def insert_data(self, data: pd.DataFrame, overwrite: bool = True) -> None:
+    def insert_data(self, data: pd.DataFrame) -> None:
         """
         Insert data into the vector database.
 
         Args:
-            data (pd.DataFrame): The data to be inserted.
-            overwrite (bool, optional): Whether to overwrite existing data. Defaults to True.
-        """
-        pass
-
-    @abstractmethod
-    def delete_data(self, filter_conditions: Dict[str, Any]) -> None:
-        """
-        Delete data from the vector database based on filter conditions.
-
-        Args:
-            filter_conditions (Dict[str, Any]): Conditions to filter which data to delete.
-        """
-        pass
-
-    @abstractmethod
-    def update_data(
-        self, filter_conditions: Dict[str, Any], new_data: Dict[str, Any]
-    ) -> None:
-        """
-        Update existing data in the vector database based on filter conditions.
-
-        Args:
-            filter_conditions (Dict[str, Any]): Conditions to filter which data to update.
-            new_data (Dict[str, Any]): New data to update the existing records with.
+            data: The data to insert.
         """
         pass
 
@@ -68,17 +32,15 @@ class BaseVectorDB(ABC):
         self,
         query_embedding: List[float],
         k: int = 10,
-        **kwargs: Any,
     ) -> List[str]:
         """
-        Perform a similarity search by vector and return results in the desired format.
+        Perform a similarity search and return matching IDs.
 
         Args:
-            query_embedding (List[float]): The embedding vector to query.
-            k (int, optional): Number of nearest neighbors to return. Defaults to 10.
-            **kwargs (Any): Additional keyword arguments for the query.
+            query_embedding: The vector to search with.
+            k: Number of nearest neighbors to return.
 
         Returns:
-            List[str]: List of result identifiers.
+            List of result IDs.
         """
         pass
