@@ -1,5 +1,4 @@
-from typing import Any, Dict, List, TypedDict
-from pathlib import Path
+from typing import List
 import pandas as pd
 import numpy as np
 from nano_vectordb import NanoVectorDB
@@ -7,8 +6,6 @@ from nano_vectordb import NanoVectorDB
 from .base_vector_db import BaseVectorDB
 
 from tigergraphx.config import NanoVectorDBConfig
-
-Data = TypedDict("Data", {"__id__": str, "__vector__": np.ndarray})
 
 
 class NanoVectorDBManager(BaseVectorDB):
@@ -21,33 +18,22 @@ class NanoVectorDBManager(BaseVectorDB):
         config: NanoVectorDBConfig,
     ):
         """
-        Initialize the NanoVectorDBWrapper.
+        Initialize the NanoVectorDBManager.
 
         Args:
-            config (NanoVectorDBConfig): Configuration for NanoVectorDB.
+            config: Configuration for NanoVectorDB.
         """
         super().__init__(config)
         self._db = NanoVectorDB(
             embedding_dim=config.embedding_dim, storage_file=str(config.storage_file)
         )
 
-    def connect(self, uri: str | Path, **kwargs: Any) -> None:
-        """
-        Connect to the vector database and set up the connection.
-
-        Args:
-            uri (str | Path): The URI or path to the database.
-            **kwargs (Any): Additional parameters for the connection.
-        """
-        pass
-
-    def insert_data(self, data: pd.DataFrame, overwrite: bool = True) -> None:
+    def insert_data(self, data: pd.DataFrame) -> None:
         """
         Insert data into NanoVectorDB.
 
         Args:
-            data (pd.DataFrame): DataFrame containing the data to insert.
-            overwrite (bool, optional): Whether to overwrite existing data. Defaults to True.
+            data: DataFrame with data to insert.
         """
         records = []
         for _, row in data.iterrows():
@@ -59,70 +45,20 @@ class NanoVectorDBManager(BaseVectorDB):
 
         self._db.upsert(records)
 
-    def delete_data(self, filter_conditions: Dict[str, Any]) -> None:
-        """
-        Delete data from NanoVectorDB based on filter conditions.
-
-        Args:
-            filter_conditions (Dict[str, Any]): Conditions to filter rows for deletion.
-        """
-
-        def filter_fn(data: Data) -> bool:
-            return all(
-                data.get(key) == value for key, value in filter_conditions.items()
-            )
-
-        ids_to_delete = [
-            data["__id__"]
-            for data in self._db.query(
-                query=np.zeros(self._db.embedding_dim), filter_lambda=filter_fn
-            )
-        ]
-        self._db.delete(ids_to_delete)
-
-    def update_data(
-        self, filter_conditions: Dict[str, Any], new_data: Dict[str, Any]
-    ) -> None:
-        """
-        Update existing data in NanoVectorDB.
-
-        Args:
-            filter_conditions (Dict[str, Any]): Conditions to filter rows for updating.
-            new_data (Dict[str, Any]): New data to update the filtered rows with.
-        """
-
-        def filter_fn(data: Data) -> bool:
-            return all(
-                data.get(key) == value for key, value in filter_conditions.items()
-            )
-
-        records_to_update = self._db.query(
-            query=np.zeros(self._db.embedding_dim), filter_lambda=filter_fn
-        )
-
-        updated_records = []
-        for record in records_to_update:
-            record.update(new_data)
-            updated_records.append(record)
-
-        self._db.upsert(updated_records)
-
     def query(
         self,
         query_embedding: List[float],
         k: int = 10,
-        **kwargs: Any,
     ) -> List[str]:
         """
-        Perform a similarity search and return results.
+        Perform a similarity search and return the result IDs.
 
         Args:
-            query_embedding (List[float]): Query embedding vector for similarity search.
-            k (int, optional): Number of top results to retrieve. Defaults to 10.
-            **kwargs (Any): Additional parameters for the query.
+            query_embedding: Embedding vector for search.
+            k: Number of top results to retrieve.
 
         Returns:
-            List[str]: List of IDs from the search results.
+            List of IDs from the search results.
         """
         results = self._db.query(query=np.array(query_embedding), top_k=k)
         return [result["__id__"] for result in results]
